+++ Console/Thrusters.h
@@ -8,11 +8,11 @@ private:
                                 {-0.35360679,  0.35360679 , 0.25879917},
                                 {0.35360679, -0.35360679 , 0.25879917},
                                 {0.35360679  ,0.35360679 ,-0.25879917}};
-    double thruster_frame[6]={0}; //UR , BR , UL ,BL , UP , DOWN
+    int thruster_frame[6]={0}; //UR , BR , UL ,BL , UP , DOWN
 public:
     void set_h_forces(int *array);
     void set_v_forces(int *array);
-    double * get_thruster_frame();
+    int * get_thruster_frame();
 
 };
